Added extra prints to content building. Fixed a problem with file dates in Emscripten. Made content always build during save (no selecting building). Moved web to using single file building. Need to fix operator new with Emscripten build by rebuilding external libraries. Made the renderer stop checking compile/link state (unless in debug).

+++ Systems/Engine/CogId.cpp
@@ -51,9 +51,9 @@ u64 CogId::ToUint64() const
   return  *(u64*)this;
 }
 
-u32 CogId::Hash() const
+size_t CogId::Hash() const
 {
-  return Id ^ Slot;
+  return (size_t)(Id ^ Slot);
 }
 
 bool CogId::IsValid()
