+++ Systems/Engine/Configuration.cpp
@@ -113,7 +113,7 @@ void UserConfig::Serialize(Serializer& stream)
 
   SerializeNameDefault(LastVersionKnown, 0u);
   SerializeNameDefault(LastVersionUsed, 0u);
-  SerializeNameDefault(LastAcceptedEulaHash, (size_t)0);
+  SerializeNameDefault(LastAcceptedEulaHash, (u64)0);
 }
 
 //------------------------------------------------------------- Developer Config
